requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.2.0
pydantic==2.6.1

# YouTube API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Audio processing
openai-whisper==20231117
ffmpeg-python==0.2.0

# Vector storage
chromadb==0.4.22
sqlalchemy==2.0.27

# PDF and document processing
pypdf==4.0.1
gspread==5.12.0
oauth2client==4.1.3

# DeepSeek API integration
# deepseek-api==0.1.0  # Replace with actual package if different
requests==2.31.0  # For API calls

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
pytest==7.4.3
pytest-cov==4.1.0 